% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aft_based.R
\name{aft_based}
\alias{aft_based}
\title{Confidence interval based on AFT model}
\usage{
aft_based(
  x,
  c,
  alpha,
  data_fit,
  data_calib,
  dist = "weibull",
  weight_calib,
  weight_new
)
}
\arguments{
\item{x}{a vector of the covariate of the test data.}

\item{c}{the censoring time cutoff.}

\item{alpha}{a number betweeo 0 and 1, specifying the miscaverage rate.}

\item{data_fit}{a data frame, containing the training data.}

\item{data_calib}{a data frame, containing the calibration data.}

\item{dist}{The distribution of T used in the aft model (default: weibull).}

\item{weight_calib}{The weight corresponding to the calibration data.}

\item{weight_new}{The weight corresponding to the test data.}
}
\value{
low_ci a value of the lower bound for the survival time of the test point.

includeR 0 or 1, indicating if [r,inf) is included in the confidence interval.
}
\description{
Construct conformal predictive interval based on the accelareted failure-time (AFT) model
}
\seealso{
Other model: 
\code{\link{distBoost_based}()},
\code{\link{gpr_based}()},
\code{\link{np_based}()},
\code{\link{ph_based}()},
\code{\link{portnoy_based}()},
\code{\link{pow_based}()},
\code{\link{quantBoost_based}()},
\code{\link{rf_based}()}
}
\concept{model}
