% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfsurv.R
\name{cfsurv}
\alias{cfsurv}
\title{Predictive confidence interval for survival data}
\usage{
cfsurv(
  x,
  c_list = NULL,
  pr_list = NULL,
  pr_new_list = NULL,
  Xtrain,
  C,
  event,
  time,
  alpha = 0.05,
  type = "quantile",
  seed = 24601,
  model = "cox",
  dist = "weibull",
  I_fit = NULL,
  ftol = 0.1,
  tol = 0.1,
  n.tree = 500
)
}
\arguments{
\item{x}{a vector of the covariate for test point.}

\item{c_list}{the cutoff for the censoring time (a positive number or a vector).}

\item{pr_list}{the censoring probability at the cutoff for the fitting and calibration datasets.}

\item{pr_new_list}{the censoring probability at the cutoff for the test points.}

\item{Xtrain}{a n-by-p matrix of the covariate of the training data.}

\item{C}{a length n vector of the censoring time of the training data.}

\item{event}{a length n vector of indicators if the time observed is censored. TRUE corresponds to NOT censored, and FALSE censored.}

\item{time}{a vevtor of length n, containing the observed survival time.}

\item{alpha}{a number between 0 and 1, speciifying the miscoverage rate.}

\item{seed}{an integer random seed (default: 24601).}

\item{model}{Options include "cox", "aft", "randomforest", "Powell", "Portnoy", "PengHuang", "distBoost", "quantBoost", "gpr". This determines the model used to fit the conformal score (default: "cox").}

\item{dist}{either "weibull", "exponential" or "gaussian" (default: "weibull"). The distribution of T used in the AFT model.}

\item{I_fit}{The set of indices corresponding to the fitting data set. (Default: NULL)}
}
\description{
The main function to generate a predictive conformal confidence interval for a unit's survival time.
}
\examples{
# Generate data
n <- 500
X <- runif(n,0,2)
T <- exp(X+rnorm(n,0,1))
C <- rexp(n,rate = 0.01)
event <- (T <= C)
time <- pmin(T,C)
data <- data.frame(X = X, C = C, event = event, censored_T = censored_T)

# Prediction point
x <- seq(0,2,by=.4)
r <- 2

# Run cfsurv
res <- cfsurv(x,r,X,R,event,time,alpha=0.1,model="cox")

}
