% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_survival.R
\name{CFsurvival.nuisance.options}
\alias{CFsurvival.nuisance.options}
\title{Initialize options for nuisance estimation}
\usage{
CFsurvival.nuisance.options(
  cross.fit = TRUE,
  V = ifelse(cross.fit, 10, 1),
  folds = NULL,
  eval.times = NULL,
  event.SL.library = lapply(c("survSL.km", "survSL.coxph", "survSL.expreg",
    "survSL.weibreg", "survSL.loglogreg", "survSL.gam", "survSL.rfsrc"), function(alg)
    c(alg, "survscreen.glmnet", "survscreen.marg", "All")),
  event.pred.0 = NULL,
  event.pred.1 = NULL,
  cens.SL.library = lapply(c("survSL.km", "survSL.coxph", "survSL.expreg",
    "survSL.weibreg", "survSL.loglogreg", "survSL.gam", "survSL.rfsrc"), function(alg)
    c(alg, "survscreen.glmnet", "survscreen.marg", "All")),
  cens.trunc = 0,
  cens.pred.0 = NULL,
  cens.pred.1 = NULL,
  survSL.control = list(initWeightAlg = "survSL.rfsrc", verbose = FALSE),
  survSL.cvControl = list(V = 10),
  save.nuis.fits = FALSE,
  prop.SL.library = lapply(c("SL.mean", "SL.glm", "SL.gam", "SL.earth", "SL.ranger",
    "SL.xgboost"), function(alg) c(alg, "screen.glmnet", "screen.corRank", "All")),
  prop.trunc = 0,
  prop.pred = NULL,
  verbose = FALSE
)
}
\arguments{
\item{cross.fit}{Logical indicating whether to cross-fit nuisance parameters. Defaults to \code{TRUE}.}

\item{V}{Positive integer number of folds for cross-fitting. Defaults to 10.}

\item{folds}{Optional \code{n x 1} vector indicating which fold each observation is in. If \code{NULL}, folds will be randomly assigned in such a way to balance observed events across folds.}

\item{eval.times}{Grid of time values on which to perform the estimation procedure. Defaults to the sorted unique values in \code{time}.}

\item{event.SL.library}{The library of candidate learners to be passed on to \code{\link[survSuperLearner]{survSuperLearner}} to estimate the conditional survival of the event. If only a single learner is provided(e.g. \code{survSL.km} for Kaplan-Meier estimator, \code{survSL.coxph} for Cox model, or \code{survSL.rfsrc} for survival random forest), then just this learner will be used, and no super learning will be performed. Defaults to a full library with screening and all algorithms currently implemented in \code{survSuperLearner}. If \code{event.SL.library = NULL}, then \code{event.pred.0} and/or \code{event.pred.1} must be specified.}

\item{event.pred.0}{Optional \code{n x k} matrix of estimates of the conditional survival of the event given treatment = 0 and confounders. If \code{event.SL.library = NULL} and \code{0 \%in\% fit.treat}, then \code{event.pred.0} must be specified. If \code{event.SL.library} is not \code{NULL}, then \code{event.pred.0 } is ignored.}

\item{event.pred.1}{Optional \code{n x k} matrix of estimates of the conditional survival of the event given treatment = 1 and confounders. If \code{event.SL.library = NULL} and \code{1 \%in\% fit.treat}, then \code{event.pred.1} must be specified. If \code{event.SL.library} is not \code{NULL}, then \code{event.pred.1 } is ignored.}

\item{cens.SL.library}{The library of candidate learners to estimate the conditional survival of censoring. As with \code{event.SL.library}, single learners can be specified. Defaults to a full library with screening and all algorithms currently implemented in \code{survSuperLearner}. If \code{cens.SL.library = NULL}, then \code{cens.pred.0} and/or \code{cens.pred.1} must be specified.}

\item{cens.trunc}{Optional lower truncation for censoring probabilities. Any estimated probabilities less than \code{cens.trunc} will be set to \code{cens.trunc}. Defaults to zero, i.e. no truncation.}

\item{cens.pred.0}{Optional \code{n x k} matrix of estimates of the conditional survival of censoring given treatment = 0 and confounders. If \code{cens.SL.library = NULL} and \code{0 \%in\% fit.treat}, then \code{cens.pred.0} must be specified. If \code{cens.SL.library} is not \code{NULL}, then \code{cens.pred.0 } is ignored.}

\item{cens.pred.1}{Optional \code{n x k} matrix of estimates of the conditional survival of censoring given treatment = 1 and confounders. If \code{cens.SL.library = NULL} and \code{1 \%in\% fit.treat}, then \code{cens.pred.1} must be specified. If \code{cens.SL.library} is not \code{NULL}, then \code{cens.pred.1} is ignored.}

\item{survSL.control}{Optional list of controls to be passed to \code{\link[survSuperLearner]{survSuperLearner}}. See \code{\link[survSuperLearner]{survSuperLearner.control}}.}

\item{survSL.cvControl}{Optional list of controls to be passed to \code{\link[survSuperLearner]{survSuperLearner}}. See \code{\link[survSuperLearner]{survSuperLearner.CV.control}}.}

\item{save.nuis.fits}{Logical indicating whether to save the fitted nuisance objects.}

\item{prop.SL.library}{The library to use for estimation of the treatment propensities using \code{\link[SuperLearner]{SuperLearner}}. If only a single learner is provided(e.g. \code{SL.mean} for marginal mean, \code{SL.glm} for logistic regression, or \code{SL.ranger} for random forest), then just this learner will be used, and no super learning will be performed. If \code{prop.SL.library = NULL}, then \code{prop.pred} must be provided.}

\item{prop.trunc}{Optional lower truncation for propensities. If 0 is in \code{fit.treat}, then any estimated propensities greater than \code{1-prop.trunc} will be set to \code{1-prop.trunc}. If 1 is in \code{fit.treat}, then any estimated propensities less than \code{prop.trunc} will be set to \code{prop.trunc}. Defaults to zero, i.e. no truncation.}

\item{prop.pred}{Optional \code{n x 1} numeric vector of estimated probabilities that \code{treat = 1} given the confounders. If \code{prop.SL.library = NULL}, then \code{prop.pred} must be specified, otherwise it is ignored.}
}
\value{
Named list containing the nuisance options.
}
\description{
This function initializes the options for nuisance parameter estimation (i.e. conditional survivals of event and censoring and treatment propensity) for use in \code{CFsurvival}. The conditional survivals of event and censoring are estimated using \code{\link[survSuperLearner]{survSuperLearner}}, or any of the individual learners therein. Treatment propensity is estimated using the \code{\link[SuperLearner]{SuperLearner}}, or any of the individual learners therein. Alternatively, the estimation process can be overriden by providing predictions from pre-fit nuisance estimators.
}
