% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_based.R
\name{cox_based}
\alias{cox_based}
\title{cox_based}
\usage{
cox_based(x, r, alpha, data_fit, data_calib, type, dist, h)
}
\arguments{
\item{x}{the covariate of the prediction point}

\item{r}{the observing time of the prediction point}

\item{alpha}{level of confidence. The target is to contruct a level 1-alpha confidence interval}

\item{data_fit}{traing data}

\item{data_calib}{calibration data}

\item{type}{type of confidence interval: marginal or local}

\item{dist}{distribution of T}
}
\description{
Construct conformal prediction model based on cox model
}
