% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_surv_prob}
\alias{extract_surv_prob}
\title{Function to extract P(T>t) [to be changed to P(T>=t)]}
\usage{
extract_surv_prob(t, time, surv)
}
\description{
Function to extract P(T>t) [to be changed to P(T>=t)]
}
