% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit the model to estimate the quantile of survival time}
\usage{
fit_model(data, model = "cox", dist = "weibull")
}
\arguments{
\item{data}{data frame used to fit the quantile regression model. It should contain (X,R,censoredT)}

\item{model}{the model used to fit the quantile regression. Takes value in ("cox","randomforest"). The default is "cox".}
}
\value{
The fitted model
}
\description{
Fit the model to estimate the quantile of survival time
}
