% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_weibull}
\alias{generate_weibull}
\title{Generate survival data generated from a cox model}
\usage{
generate_weibull(n, a, b, beta, Rmax, seed, info = FALSE)
}
\description{
Generate survival data generated from a cox model
}
\keyword{internal}
