% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score.R
\name{get_calibration}
\alias{get_calibration}
\title{Computing the calibration term with covaraite shift}
\usage{
get_calibration(score, weight_calib, weight_new, alpha)
}
\arguments{
\item{alpha}{a number betweeo 0 and 1, specifying the miscaverage rate.}

\item{x}{a vector of the covariate of the test data.}

\item{r}{the censoring time of the test data.}

\item{data}{a data frame used for calibration, containing four columns: (X,R,event,censored_T).}

\item{mdl}{The fitted model to estimate the conditional quantile (default is NULL).}

\item{quant_lo}{the fitted conditional quantile for the calibration data (default is NULL).}

\item{new_quant_lo}{the fitted conditional quantile for the test data (default is NULL).}
}
\value{
low_ci a value of the lower bound for the survival time of the test point.

includeR 0 or 1, indicating if [r,inf) is included in the confidence interval.
}
\description{
construct the one-sided confidence interval for a unit's survival time T
}
\seealso{
Other confint: 
\code{\link{lower_ci_local}()},
\code{\link{lower_ci}()},
\code{\link{scaled_lower_ci}()}
}
\concept{confint}
