% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_score.R
\name{lower_score}
\alias{lower_score}
\title{lower_score}
\usage{
lower_score(mdl, data, alpha)
}
\arguments{
\item{data}{calibration data frame}

\item{alpha}{1-alpha is the level of the confidence interval}

\item{model}{The fitted model}
}
\description{
lower_score
}
