% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_based.R
\name{np_based}
\alias{np_based}
\title{Confidence interval based on nonparametric regression}
\usage{
np_based(
  x,
  c,
  alpha,
  data_fit,
  data_calib,
  weight_calib,
  weight_new,
  ftol = 0.1,
  tol = 0.2
)
}
\arguments{
\item{x}{a vector of the covariate of the test data.}

\item{c}{the censoring time cutoff.}

\item{alpha}{a number betweeo 0 and 1, specifying the miscaverage rate.}

\item{data_fit}{a data frame, containing the training data.}

\item{data_calib}{a data frame, containing the calibration data.}

\item{weight_calib}{The weight corresponding to the calibration data.}

\item{weight_new}{The weight corresponding to the test data.}
}
\value{
low_ci a value of the lower bound for the survival time of the test point.

includeR 0 or 1, indicating if [r,inf) is included in the confidence interval.
}
\description{
Construct conformal predictive interval based on nonparametric regression
}
\seealso{
Other model: 
\code{\link{aft_based}()},
\code{\link{distBoost_based}()},
\code{\link{gpr_based}()},
\code{\link{ph_based}()},
\code{\link{portnoy_based}()},
\code{\link{pow_based}()},
\code{\link{quantBoost_based}()},
\code{\link{rf_based}()}
}
\concept{model}
