% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_res}
\alias{plot_res}
\title{plot_res}
\usage{
plot_res(
  x_base,
  r,
  alpha_list,
  upp_ci = NULL,
  low_ci,
  method_name,
  lower_only = FALSE,
  includeR = NULL,
  dir,
  plot
)
}
\description{
An internal function to plot the result
}
