% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portnoy_based.R
\name{portnoy_based}
\alias{portnoy_based}
\title{Confidence interval based on Portnoy (2003)}
\usage{
portnoy_based(x, r, alpha, data_fit, data_calib, type, h)
}
\arguments{
\item{x}{a vector of the covariate of the test data.}

\item{r}{the censoring time of the test data.}

\item{alpha}{a number betweeo 0 and 1, specifying the miscaverage rate.}

\item{data_fit}{a data frame, containing the training data.}

\item{data_calib}{a data frame, containing the calibration data.}

\item{type}{either "marginal" or "local". Determines the type of confidence interval.}
}
\value{
low_ci a value of the lower bound for the survival time of the test point.

includeR 0 or 1, indicating if [r,inf) is included in the confidence interval.
}
\description{
Construct conformal predictive interval based on cox model
}
\seealso{
Other model: 
\code{\link{cox_based}()},
\code{\link{ph_based}()},
\code{\link{pow_based}()},
\code{\link{rf_based}()}
}
\concept{model}
