% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powell_based.R
\name{pow_based}
\alias{pow_based}
\title{Confidence interval based on Powell (1986)}
\usage{
pow_based(x, c, alpha, data_fit, data_calib, weight_calib, weight_new)
}
\arguments{
\item{x}{a vector of the covariate of the test data.}

\item{alpha}{a number betweeo 0 and 1, specifying the miscaverage rate.}

\item{data_fit}{a data frame, containing the training data.}

\item{data_calib}{a data frame, containing the calibration data.}

\item{r}{the censoring time of the test data.}

\item{type}{either "marginal" or "local". Determines the type of confidence interval.}
}
\value{
low_ci a value of the lower bound for the survival time of the test point.

includeR 0 or 1, indicating if [r,inf) is included in the confidence interval.
}
\description{
Construct conformal predictive interval based on cox model
}
\seealso{
Other model: 
\code{\link{cox_based}()},
\code{\link{distBoost_based}()},
\code{\link{gpr_based}()},
\code{\link{ph_based}()},
\code{\link{portnoy_based}()},
\code{\link{quantBoost_based}()},
\code{\link{rf_based}()}
}
\concept{model}
