% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_km.R
\name{ranger.getWeights}
\alias{ranger.getWeights}
\title{ranger.getWeights}
\usage{
ranger.getWeights(rg, traindata, testdata, y_name, c_name, x_name = NULL)
}
\description{
An internal function for censored quantile random forest
}
