% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_based.R
\name{rf_based}
\alias{rf_based}
\title{rf_based}
\usage{
rf_based(x, r, alpha, data_fit, data_calib, type, h)
}
\arguments{
\item{x}{covariate of the prediction point}
}
\description{
Construct a conformal prediction confidence interval based on random forest
}
