% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_scaled.R
\name{rf_scaled}
\alias{rf_scaled}
\title{Confidence interval based on random forest (scaled)}
\usage{
rf_scaled(x, r, alpha, data_fit, data_calib, type, h)
}
\arguments{
\item{x}{a vector of the covariate of the test data.}

\item{r}{the censoring time of the test data.}

\item{alpha}{a number betweeo 0 and 1, specifying the miscaverage rate.}

\item{data_fit}{a data frame, containing the training data.}

\item{data_calib}{a data frame, containing the calibration data.}

\item{type}{either "marginal" or "local". Determines the type of confidence interval.}

\item{dist}{The distribution of T used in the cox model.}
}
\value{
low_ci a value of the lower bound for the survival time of the test point.

includeR 0 or 1, indicating if [r,inf) is included in the confidence interval.
}
\description{
Construct a conformal predictive confidence interval based on random forest
}
\seealso{
Other model: 
\code{\link{adaptive_cox_based}()},
\code{\link{cox_based}()},
\code{\link{cox_scaled}()},
\code{\link{ph_based}()},
\code{\link{portnoy_based}()},
\code{\link{pow_based}()},
\code{\link{rf_based}()}
}
\concept{model}
