% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled_lower_ci.R
\name{scaled_lower_ci}
\alias{scaled_lower_ci}
\title{One-sided predictive confidence interval (scaled)}
\usage{
scaled_lower_ci(x, r, alpha, data, quant_lo, new_quant_lo, weight, new_weight)
}
\arguments{
\item{x}{a vector of the covariate of the test data.}

\item{r}{the censoring time of the test data.}

\item{alpha}{a number betweeo 0 and 1, specifying the miscaverage rate.}

\item{data}{a data frame used for calibration, containing four columns: (X,R,event,censored_T).}

\item{quant_lo}{the fitted conditional quantile for the calibration data (default is NULL).}

\item{new_quant_lo}{the fitted conditional quantile for the test data (default is NULL).}

\item{mdl}{The fitted model to estimate the conditional quantile (default is NULL).}
}
\value{
low_ci a value of the lower bound for the survival time of the test point.

includeR 0 or 1, indicating if [r,inf) is included in the confidence interval.
}
\description{
construct the one-sided confidence interval for a unit's survival time T
}
\seealso{
Other confint: 
\code{\link{get_calibration}()},
\code{\link{lower_ci_local}()},
\code{\link{lower_ci}()}
}
\concept{confint}
