% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_c.R
\name{selection_c}
\alias{selection_c}
\title{Automatic selection of censoring time}
\usage{
selection_c(
  X,
  C,
  event,
  time,
  alpha,
  c_ref,
  weight_ref,
  model = "cox",
  type = "quantile",
  dist = "weibull"
)
}
\description{
The funciton that automatically choose a value of c_0.
}
