% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{censoring_prob}
\alias{censoring_prob}
\title{Fitting the censoring probability P(C>=c|X)}
\usage{
censoring_prob(fit, calib, test = NULL, xnames, c)
}
\description{
Fitting the censoring probability P(C>=c|X)
}
