% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfsurv.R
\name{cfsurv}
\alias{cfsurv}
\title{cfsurv}
\usage{
cfsurv(
  x,
  r,
  data,
  alpha = 0.05,
  type = "marginal",
  seed = 24601,
  model = "cox",
  dist = "weibull",
  h = 1
)
}
\arguments{
\item{x}{the covariate for prediction point. (Currently only one-dimensional.)}

\item{r}{the censoring time for the prediction point.}

\item{data}{a data frame containing the training data. It should contain four columns: (X,R,event,censoredT). X is the covariate; R is the censoring tim;, event is the indicator if T<=R; censored_T is the censored survial time, i.e., the minimum of T and R.}

\item{alpha}{The miscoverage rate.}

\item{seed}{The random seed. Default is 24601.}

\item{model}{The model used to fit the quantile. Choices include "cox" and "randomforest". Default is "cox".}

\item{dist}{The distribution of T used in the cox model. Choices include "weibull", "exponential" and "gaussian". Default is "weibull".}

\item{h}{The bandwidth for the local confidence interval. Default is 1.}
}
\description{
The main function to generate a predictive conformal confidence interval for a unit's survival time.
}
\examples{
# Generate data
n <- 500
X <- runif(n,0,2)
T <- exp(X+rnorm(n,0,1)) 
R <- rexp(n,rate = 0.01)
event <- T<=R
censored_T <- pmin(T,R)
data <- data.frame(X=X,R=R,evemt=event,censored_T=censored_T)

# Prediction point
x <- seq(0,2,by=.4)
r <- 2

# Run cfsurv
res <- cfsurv(x,r,data,alpha=0.1,model="randomforest")

}
