% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfsurv.R
\name{cfsurv}
\alias{cfsurv}
\title{Predictive confidence interval for survival data}
\usage{
cfsurv(
  x,
  c_list = NULL,
  pr_list = NULL,
  pr_new_list = NULL,
  Xtrain,
  C,
  event,
  time,
  alpha = 0.05,
  type = "quantile",
  seed = 24601,
  model = "cox",
  dist = "weibull",
  I_fit = NULL,
  ftol = 0.1,
  tol = 0.1
)
}
\arguments{
\item{x}{a vector of the covariate for test point.}

\item{Xtrain}{a n-by-p matrix of the covariate of the training data.}

\item{C}{a length n vector of the censoring time of the training data.}

\item{event}{a length n vector of indicators if the time observed is censored. TRUE corresponds to NOT censored, and FALSE censored.}

\item{time}{a vevtor of length n, containing the observed survival time.}

\item{alpha}{a number between 0 and 1, speciifying the miscoverage rate.}

\item{seed}{an integer random seed (default: 24601).}

\item{model}{Options include "cox", "randomforest", "Powell", "Portnoy" and "PengHuang". This determines the model used to fit the condditional quantile (default: "cox").}

\item{dist}{either "weibull", "exponential" or "gaussian" (default: "weibull"). The distribution of T used in the cox model.}

\item{c}{the censoring time for the test point.}

\item{h}{the bandwidth for the local confidence interval. Default is 1.}
}
\value{
low_ci a value of the lower bound for the survival time of the test point.

includeR 0 or 1, indicating if [r,inf) is included in the confidence interval.
}
\description{
The main function to generate a predictive conformal confidence interval for a unit's survival time.
}
\examples{
# Generate data
n <- 500
X <- runif(n,0,2)
T <- exp(X+rnorm(n,0,1))
R <- rexp(n,rate = 0.01)
event <- T<=R
time <- pmin(T,R)
data <- data.frame(X=X,R=R,event=event,censored_T=censored_T)
# Prediction point
x <- seq(0,2,by=.4)
r <- 2
# Run cfsurv
res <- cfsurv(x,r,X,R,event,time,alpha=0.1,model="cox")

}
