% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfsurv.R
\name{cfsurv}
\alias{cfsurv}
\title{cfsurv}
\usage{
cfsurv(
  x,
  r,
  data,
  alpha = 0.05,
  type = "marginal",
  seed = 24601,
  model = "cox",
  dist = "weibull",
  h = NULL
)
}
\arguments{
\item{x}{the covariate for prediction point}

\item{r}{the censoring time for the prediction point}

\item{data}{the data frame (X,R,censoredT)}
}
\description{
The main function to generate a predictive conformal confidence interval for a unit's survival time.
}
