% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_ci.R
\name{lower_ci}
\alias{lower_ci}
\title{lower_ci
construct the one-sided confidence interval for T}
\usage{
lower_ci(mdl, x, r, alpha, data)
}
\arguments{
\item{mdl}{The fitted model}

\item{x}{The covariate of the prediction point}

\item{r}{The observe time of the prediction point}

\item{alpha}{level}

\item{data}{data frame}
}
\description{
lower_ci
construct the one-sided confidence interval for T
}
