% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_ci.R
\name{lower_ci}
\alias{lower_ci}
\title{lower_ci}
\usage{
lower_ci(
  x,
  r,
  alpha,
  data = data,
  mdl = NULL,
  quant_lo = NULL,
  new_quant_lo = NULL
)
}
\arguments{
\item{x}{The covariate of the prediction point}

\item{r}{The observe time of the prediction point}

\item{alpha}{The miscaverage rate.}

\item{data}{A data frame used for calibration. Should contain four columns: (X,R,event,censored_T).}

\item{mdl}{The fitted model to estimate the conditional quantile. The default is NULL.}

\item{quant_lo}{The fitted conditional quantile for the calibration data.}

\item{new_quant_lo}{The fitted conditional quantile for the tprediction data.}
}
\description{
construct the one-sided confidence interval for a unit's survival time T
}
