% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_ci_local.R
\name{lower_ci_local}
\alias{lower_ci_local}
\title{lower_ci_local}
\usage{
lower_ci_local(mdl, x, r, data, alpha, h = 1)
}
\arguments{
\item{mdl}{the fitted moodel}

\item{x}{covariate of the prediction point. Can be a point or a vector}

\item{r}{observe time of the prediction point. Can be point or a vector. When it is a vector, its length should match the length of x.}

\item{data}{data frame containing the calibration data}

\item{alpha}{construct a level 1-alpha confidence interval}

\item{h}{bandwidth (default: 1)}
}
\description{
A function to compute the local lower bound of the survival time
}
