% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_quantiles}
\alias{plot_quantiles}
\title{Plot the true quantile and the conformal quantiles}
\usage{
plot_quantiles(x_base, true_lo, cox_lo, rf_lo, alpha, dir, plot = FALSE)
}
\description{
Plot the true quantile and the conformal quantiles
}
\keyword{internal}
